feat(gtm): rework state shape and card architecture per PR review round 2
- Move all GTM state under a root `google_tag_manager` key (connection/
  existing_accounts/containers), out of the shared `accounts` bucket
- Drop the stored `step` field; derive UI state from `status` alone
- Drop stored `tagManagerUrl`; build it client-side from the account ID
- Split the not-yet-connected flow into its own
  `ConnectGoogleTagManagerAccountCard`, narrowing
  `IncompleteGoogleTagManagerAccountCard` to container-selection only
- Once an account/container is picked, read display data straight off
  the connection record instead of cross-referencing the accounts/
  containers lists
- Rename typedefs: GoogleTagManagerAccountRef -> GoogleTagManagerAccount,
  GoogleTagManagerContainerRef -> GoogleTagManagerContainer,
  GoogleTagManagerAccount -> GoogleTagManagerConnection
- Simplify NoticeDetail's body prop to accept JSX directly instead of an
  array of lines
