+++ js/src/hooks/useExistingGoogleTagManagerAccounts.js
@@ -8,15 +8,24 @@ import { useSelect } from '@wordpress/data';
  */
 import { STORE_KEY } from '~/data/constants';
 
+const selectorName = 'getExistingGoogleTagManagerAccounts';
+
+/**
+ * A hook to load the Google Tag Manager accounts available to the connected Google user.
+ *
+ * @return {{ existingAccounts: Object[]|null, hasFinishedResolution: boolean }} The data and its resolution state.
+ */
 const useExistingGoogleTagManagerAccounts = () => {
 	return useSelect( ( select ) => {
-		const existingAccounts =
-			select( STORE_KEY ).getExistingGoogleTagManagerAccounts();
-		const hasFinishedResolution = select( STORE_KEY ).hasFinishedResolution(
-			'getExistingGoogleTagManagerAccounts'
-		);
+		const selector = select( STORE_KEY );
 
-		return { existingAccounts, hasFinishedResolution };
+		return {
+			existingAccounts: selector[ selectorName ](),
+			hasFinishedResolution: selector.hasFinishedResolution(
+				selectorName,
+				[]
+			),
+		};
 	}, [] );
 };
