+++ js/src/pages/settings/accounts/google-tag-manager-account-card/hooks/useConnectGoogleTagManagerAccount.js
@@ -13,11 +13,9 @@ import useDispatchCoreNotices from '~/hooks/useDispatchCoreNotices';
 
 /**
  * A hook that connects the merchant's picked Google Tag Manager account and refreshes connection
- * state. Refetches only the connection resolver (not the accounts/containers list resolvers) —
- * mirrors `useUpsertAdsAccount`'s targeted refetch-after-mutate approach, so this action doesn't
- * force every GTM consumer to re-resolve.
+ * state.
  *
- * @return {{ connect: ( accountId: string ) => Promise<void>, loading: boolean }} Click handler to wire to the "Connect" button (`accountId` is the picked `GoogleTagManagerAccountRef`'s `id`, a string), and whether a request is in flight.
+ * @return {{ connect: ( accountId: string ) => Promise<void>, loading: boolean }} `connect` — click handler to wire to the "Connect" button — and whether a request is in flight.
  */
 const useConnectGoogleTagManagerAccount = () => {
 	const { createNotice } = useDispatchCoreNotices();
@@ -28,9 +26,13 @@ const useConnectGoogleTagManagerAccount = () => {
 		method: 'POST',
 	} );
 
+	/**
+	 * @param {string} accountId The picked account's ID.
+	 */
 	const connect = async ( accountId ) => {
 		try {
-			await fetchConnect( { data: { id: accountId } } );
+			// await fetchConnect( { data: { id: accountId } } );
+			await fetchConnect( accountId );
 			await fetchGoogleTagManagerAccount();
 		} catch ( error ) {
 			createNotice(
