+++ js/src/pages/settings/disconnect-modal/constants.js
@@ -1,4 +1,4 @@
 export const ALL_ACCOUNTS = 'all-accounts';
 export const ADS_ONLY = 'ads-only';
 export const YOUTUBE_ACCOUNT = 'youtube-account';
-export const TAG_MANAGER_ACCOUNT = 'tag-manager-account';
+export const GOOGLE_TAG_MANAGER_ACCOUNT = 'google-tag-manager-account';
